log-service/requirements.txt
fastapi
uvicorn
pydantic
pymemcache
python-dotenv
aioredis
redis
fastapi-limiter